abstract flashSaleDp
