+++ assets/components/featuredProductHero/featuredProductHero.scss
@@ -1,4 +1,5 @@
-.component-flash-sale-featured-digital-pack{
+
+.component-featured-product-hero {
   position: relative;
   overflow: hidden;
   background: gu-colour(garnett-neutral-3);
@@ -29,7 +30,7 @@
     }
   }
 
-  .component-flash-sale-featured-digital-pack__countdownbox {
+  .component-featured-product-hero__countdownbox {
     border-left: 1px solid gu-colour(garnett-neutral-4);
     padding-left: 5px;
     padding-bottom: 120px;
@@ -39,7 +40,7 @@
     }
   }
 
-  .component-flash-sale-featured-digital-pack__countdownbox .component-flash-sale-countdown {
+  .component-featured-product-hero__countdownbox .component-flash-sale-countdown {
     margin: 0 -5px;
     color: #bb3b80;
     width: 130%;
@@ -54,12 +55,12 @@
     }
   }
 
-  .component-flash-sale-featured-digital-pack__countdownbox--hidden {
+  .component-featured-product-hero__countdownbox--hidden {
     border-left: 0;
     padding-left: 0;
   }
 
-  .component-flash-sale-featured-digital-pack__countdownbox--hidden .component-flash-sale-countdown {
+  .component-featured-product-hero__countdownbox--hidden .component-flash-sale-countdown {
     display: none;
   }
   
@@ -95,7 +96,7 @@
   }
 }
 
-.component-flash-sale-featured-digital-pack__content {
+.component-featured-product-hero__content {
   @extend .gu-content-margin;
 
   display: flex;
@@ -117,7 +118,7 @@
   }
 }
 
-.component-flash-sale-featured-digital-pack__description {
+.component-featured-product-hero__description {
   @extend .gu-content-margin;
   position: relative;
 
@@ -129,14 +130,14 @@
   }
 }
 
-.component-flash-sale-featured-digital-pack__image {
+.component-featured-product-hero__image {
   position: absolute;
   bottom: 0;
   right: 0;
   overflow: hidden;
 }
 
-.component-flash-sale-featured-digital-pack__title {
+.component-featured-product-hero__title {
   font-size: 48px;
   line-height: 48px;
   font-family: $gu-headline;
@@ -147,8 +148,8 @@
   }
 }
 
-.component-flash-sale-featured-digital-pack__heading {
-  @extend .component-flash-sale-featured-digital-pack__title;
+.component-featured-product-hero__heading {
+  @extend .component-featured-product-hero__title;
   color: gu-colour(lifestyle-garnett-main-1);
   margin-top: 8px;
   @include mq($from: phablet) {
@@ -156,15 +157,15 @@
   }
 }
 
-.component-flash-sale-featured-digital-pack__subheading{
-  @extend .component-flash-sale-featured-digital-pack__title;
+.component-featured-product-hero__subheading{
+  @extend .component-featured-product-hero__title;
   margin-bottom: 12px;
   @include mq($from: phablet) {
     max-width: 350px;
   }
 }
 
-.component-flash-sale-featured-digital-pack__copy {
+.component-featured-product-hero__copy {
   font-size: 14px;
   line-height: 1.41;
   margin-top: $gu-v-spacing / 2;
