+++ assets/pages/subscriptions-landing/subscriptionsLanding.scss
@@ -7,7 +7,7 @@
 
 @import '~components/ctaLink/ctaLink';
 @import '~components/doubleHeading/doubleHeading';
-@import '~components/featuredDigitalPack/flashSaleDigitalPack';
+@import '~components/featuredProductHero/featuredProductHero';
 @import '~components/featureList/featureList';
 @import '~components/footer/footer';
 @import '~components/headers/simpleHeader/simpleHeader';
